+++ app/views/sessions/new.html.erb
@@ -1,4 +1,4 @@
-<div style="width: 350px; margin: 0 auto; margin-top: 50px;">
+<div style="width: 500px; margin: 0 auto; margin-top: 50px;">
   <div class="card">
     <div class="card-header">
       <h5>Login</h5>
@@ -16,10 +16,13 @@
         <br>
 
         <div class="row">
-          <div class="col-6">
+          <div class="col-4">
             <%= form.submit "Sign In" %>
           </div>
-          <div class="col-6">
+          <div class="col-4">
+            <%= link_to "Forgot password", new_password_path %>
+          </div>
+          <div class="col-4">
             <%= link_to "Sign up here", sign_up_path %>
           </div>
         </div>
