+++ app/views/shared/_navbar.html.erb
@@ -6,15 +6,19 @@
     </button>
     <div class="collapse navbar-collapse" id="mynavbar">
       <ul class="navbar-nav me-auto">
-        <li class="nav-item">
-          <a class="nav-link" href="javascript:void(0)">Link</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="javascript:void(0)">Link</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="javascript:void(0)">login</a>
-        </li>
+        <% if user_signed_in?  %>
+          <li class="nav-link">
+            <%= link_to 'Logout', logout_url %>
+          </li>
+          <%else %>
+          <li class="nav-link">
+            <%= link_to 'Login', login_path %>
+          </li>
+          <li class="nav-link">
+            <%= link_to 'Sign up', sign_up_path %>
+          </li>
+          <%end%>
+
       </ul>
       <form class="d-flex">
         <input class="form-control me-2" type="text" placeholder="Search">
